+++ changelog.md
@@ -1,2 +1,3 @@
 - Fixed gas station synchronization issues
 - Fixed top gas station block translation
+- Fixed being able to take out fluid from trading gas stations